requirements.txt
fastapi
fastembed
pillow
uvicorn